revert sul int->long. portato la compilazione a 64bit

+++ Program/LoadFile.cs
@@ -8,8 +8,6 @@ namespace Program
 {
     public class LoadFile
     {
-
-
         public Data LoadFrom(string path)
         {
             Data data = new Data();
@@ -22,8 +20,7 @@ namespace Program
             data.NumberOfEndpoints = int.Parse(li[1]);
             data.NumberOfRequests = int.Parse(li[2]);
             data.NumberOfCacheServers = int.Parse(li[3]);
-            data.CapacityOfCacheServer = long.Parse(li[4]);
-
+            data.CapacityOfCacheServer = int.Parse(li[4]);
 
             string line2 = file.ReadLine();// L2
             string[] vids = line2.Split(' ');
@@ -33,20 +30,16 @@ namespace Program
                 data.Videos.Add(v);
             }
 
-
             for (int i = 0; i < data.NumberOfEndpoints; i++)
             {
-
                 string lineEp = file.ReadLine();// EndPoint
                 string[] endP = lineEp.Split(' ');
-
 
                 EndPoint ep = new EndPoint();
                 ep.ID = i;
                 ep.LatencyDataCenter = int.Parse(endP[0]);
 
                 int connectedCaches = int.Parse(endP[1]);
-
 
                 for (int j = 0; j < connectedCaches; j++)
                 {
@@ -57,13 +50,10 @@ namespace Program
                     conn.CacheServerID = int.Parse(cc[0]);
                     conn.LatencyCache = int.Parse(cc[1]);
 
-
                     ep.ConnectedServers.Add(conn);
                 }
 
                 data.Endpoint.Add(ep);
-
-
             }
 
             for (int k = 0; k < data.NumberOfRequests; k++)
@@ -76,20 +66,12 @@ namespace Program
                 r.EndPointID = int.Parse(crrc[1]);
                 r.NumberOfRequests = int.Parse(crrc[2]);
 
-
                 data.Requests.Add(r);
             }
 
-
-
             file.Close();
-
 
             return data;
         }
-
-
-
-
     }
 }
